refactor: Crew 및 Squad 간의 관계를 Member 기준으로 반정규화
Squad와 SquadMember가 CrewMember를 통해 Member 정보를 참조하던 다단계 조인 구조를 개선함.

- 원인: Squad.creator 정보 조회 시 CrewMember를 거치는 불필요한 4단계 조인 발생.
- 변경: Squad 및 SquadMember가 Member와 직접 연관되도록 반정규화함. (스쿼드와 관련된 모든 도메인 SquadRequest, SquadMember, SquadComment 모두 CrewMember -> Member로 평탄화)
- 결과: 작성자 정보 조회 성능 향상 및 '스쿼드 참여 여부' 확인 쿼리 단순화.
- 주의: 비즈니스 규칙 검증(멤버십 확인)은 Application Service 레이어에서 수동으로 처리하도록 책임이 이동됨.

* 추가적으로. 추후 공통메서드들을 한번에 클래스로 분리하기위해 현재 중복 코드가 상당히 많음.
* 테스트는 모두 깨진 상태.

+++ src/main/java/revi1337/onsquad/squad/application/SquadCommandService.java
@@ -1,48 +1,67 @@
 package revi1337.onsquad.squad.application;
 
-import static revi1337.onsquad.squad.error.SquadErrorCode.UNSUFFICIENT_AUTHORITY;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import revi1337.onsquad.category.domain.entity.Category;
 import revi1337.onsquad.crew_member.domain.entity.CrewMember;
 import revi1337.onsquad.crew_member.domain.repository.CrewMemberRepository;
+import revi1337.onsquad.crew_member.error.CrewMemberErrorCode;
+import revi1337.onsquad.crew_member.error.exception.CrewMemberBusinessException;
 import revi1337.onsquad.squad.application.dto.SquadCreateDto;
 import revi1337.onsquad.squad.domain.entity.Squad;
 import revi1337.onsquad.squad.domain.repository.SquadRepository;
+import revi1337.onsquad.squad.error.SquadErrorCode;
 import revi1337.onsquad.squad.error.exception.SquadBusinessException;
 import revi1337.onsquad.squad_category.domain.repository.SquadCategoryJdbcRepository;
 import revi1337.onsquad.squad_member.domain.entity.SquadMember;
 import revi1337.onsquad.squad_member.domain.repository.SquadMemberRepository;
+import revi1337.onsquad.squad_member.error.SquadMemberErrorCode;
+import revi1337.onsquad.squad_member.error.exception.SquadMemberBusinessException.NotParticipant;
 
 @Transactional
 @RequiredArgsConstructor
 @Service
-public class SquadCommandService {
+public class SquadCommandService {  // TODO ApplicationService 와 DomainService 로 나눌 수 있을듯? 일단 그건 나중에
 
     private final SquadRepository squadRepository;
     private final CrewMemberRepository crewMemberRepository;
     private final SquadMemberRepository squadMemberRepository;
     private final SquadCategoryJdbcRepository squadCategoryJdbcRepository;
 
-    public Long newSquad(Long memberId, Long crewId, SquadCreateDto dto) {
-        CrewMember crewMember = crewMemberRepository.getByCrewIdAndMemberId(crewId, memberId);
-        Squad persistSquad = squadRepository.save(
-                Squad.create(dto.toEntityMetadata(), crewMember, crewMember.getCrew())
-        );
-        squadCategoryJdbcRepository.batchInsert(persistSquad.getId(), Category.fromCategoryTypes(dto.categories()));
-
-        return persistSquad.getId();
+    public void newSquad(Long memberId, Long crewId, SquadCreateDto dto) {
+        CrewMember crewMember = validateMemberInCrewAndGet(memberId, crewId);
+        Squad squad = squadRepository.save(Squad.create(dto.toEntityMetadata(), crewMember.getMember(), crewMember.getCrew()));
+        squadCategoryJdbcRepository.batchInsert(squad.getId(), Category.fromCategoryTypes(dto.categories()));
     }
 
     public void deleteSquad(Long memberId, Long crewId, Long squadId) {
-        CrewMember crewMember = crewMemberRepository.getByCrewIdAndMemberId(crewId, memberId);
-        SquadMember squadMember = squadMemberRepository.getBySquadIdAndCrewMemberId(squadId, crewMember.getId());
-        if (squadMember.isNotLeader()) {
-            throw new SquadBusinessException.CantDelete(UNSUFFICIENT_AUTHORITY);
+        Squad squad = validateSquadExistsAndGet(squadId);
+        if (squad.mismatchCrewId(crewId)) {
+            throw new SquadBusinessException.MismatchReference(SquadErrorCode.MISMATCH_CREW_REFERENCE);
         }
-
+        if (squad.mismatchMemberId(memberId)) {
+            throw new SquadBusinessException.MismatchReference(SquadErrorCode.MISMATCH_MEMBER_REFERENCE);
+        }
+        SquadMember squadMember = validateMemberInSquadAndGet(memberId, squadId);
+        if (squadMember.isNotLeader()) {
+            throw new SquadBusinessException.InsufficientAuthority(SquadErrorCode.INSUFFICIENT_DELETE_AUTHORITY);
+        }
         squadRepository.deleteById(squadId);
+    }
+
+    private CrewMember validateMemberInCrewAndGet(Long memberId, Long crewId) {  // TODO 리팩토링 싹다끝내면, 하위 private 메서드 모두 책임 분리 필요.
+        return crewMemberRepository.findByCrewIdAndMemberId(crewId, memberId)
+                .orElseThrow(() -> new CrewMemberBusinessException.NotParticipant(CrewMemberErrorCode.NOT_PARTICIPANT));
+    }
+
+    private Squad validateSquadExistsAndGet(Long squadId) {
+        return squadRepository.findById(squadId)
+                .orElseThrow(() -> new SquadBusinessException.NotFound(SquadErrorCode.NOT_FOUND));
+    }
+
+    private SquadMember validateMemberInSquadAndGet(Long memberId, Long squadId) {
+        return squadMemberRepository.findBySquadIdAndMemberId(squadId, memberId)
+                .orElseThrow(() -> new NotParticipant(SquadMemberErrorCode.NOT_PARTICIPANT));
     }
 }
